requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
boto3
requests
slowapi
python-multipart
psycopg2-binary
email-validator
argon2-cffi
resend